docs(tests): add Spanish classification comments to 1,333 tests
- Add // Clasificación: determinística — {reason} to all test functions
- Add // Clasificación: frágil — {reason} for fragile tests
- Comments explain WHY each test exists and what edge case it covers
- All comments in Spanish, all code in English
- No test logic changed — only comments added

+++ fc-examples/src/adapters/gpu_renderer.rs
@@ -579,8 +579,8 @@ impl GpuRenderer {
             .time_series
             .iter()
             .filter(|b| {
-                b.timestamp >= viewport.time_start as u64
-                    && b.timestamp <= viewport.time_end as u64
+                b.timestamp >= viewport.time_start
+                    && b.timestamp <= viewport.time_end
             })
             .copied()
             .collect();
@@ -654,8 +654,8 @@ impl GpuRenderer {
             .time_series
             .iter()
             .filter(|b| {
-                b.timestamp >= viewport.time_start as u64
-                    && b.timestamp <= viewport.time_end as u64
+                b.timestamp >= viewport.time_start
+                    && b.timestamp <= viewport.time_end
             })
             .copied()
             .collect();
@@ -729,8 +729,8 @@ impl GpuRenderer {
             .time_series
             .iter()
             .filter(|b| {
-                b.timestamp >= viewport.time_start as u64
-                    && b.timestamp <= viewport.time_end as u64
+                b.timestamp >= viewport.time_start
+                    && b.timestamp <= viewport.time_end
             })
             .copied()
             .collect();
